move bestuursperiod models up one route
