+++ app/routes/organen/orgaan.js
@@ -1,15 +1,25 @@
 import Route from '@ember/routing/route';
 import { inject as service } from '@ember/service';
+import {
+  getBestuursPeriodsOG,
+  getSelectedBestuursorgaanWithPeriods,
+} from 'frontend-lmb/utils/bestuursperioden';
 import { getFormFrom } from 'frontend-lmb/utils/get-form';
 import { BESTUURSORGAAN_FORM_ID } from 'frontend-lmb/utils/well-known-ids';
 import RSVP from 'rsvp';
 
 export default class OrganenOrgaanRoute extends Route {
   @service store;
+
+  queryParams = {
+    startDate: { refreshModel: true },
+    endDate: { refreshModel: true },
+  };
+
   async model(params) {
     const bestuursorgaanId = params.orgaan_id;
 
-    const bestuursorgaan = this.store.findRecord(
+    const bestuursorgaan = await this.store.findRecord(
       'bestuursorgaan',
       bestuursorgaanId,
       {
@@ -17,12 +27,30 @@ export default class OrganenOrgaanRoute extends Route {
       }
     );
 
+    const tijdsspecialisaties = await bestuursorgaan.heeftTijdsspecialisaties;
+    const currentBestuursOrgaanWithPeriods =
+      getSelectedBestuursorgaanWithPeriods(tijdsspecialisaties, {
+        startDate: params.startDate,
+        endDate: params.endDate,
+      });
+
+    const bestuursPeriods = getBestuursPeriodsOG(tijdsspecialisaties);
+    const selectedPeriod = {
+      startDate: currentBestuursOrgaanWithPeriods.startDate,
+      endDate: currentBestuursOrgaanWithPeriods.endDate,
+    };
+    const currentBestuursOrgaan =
+      currentBestuursOrgaanWithPeriods.bestuursOrgaan;
+
     const formDefinition = getFormFrom(this.store, BESTUURSORGAAN_FORM_ID);
 
     return RSVP.hash({
       form: formDefinition,
       instanceId: bestuursorgaanId,
       bestuursorgaan,
+      bestuursPeriods,
+      selectedPeriod,
+      currentBestuursOrgaan,
     });
   }
 }
