+++ app/routes/organen/orgaan/mandatarissen.js
@@ -1,17 +1,11 @@
 import Route from '@ember/routing/route';
 import { action } from '@ember/object';
 import { inject as service } from '@ember/service';
-import {
-  getBestuursPeriodsOG,
-  getSelectedBestuursorgaanWithPeriods,
-} from 'frontend-lmb/utils/bestuursperioden';
 
 export default class OrganenMandatarissenRout extends Route {
   @service store;
 
   queryParams = {
-    startDate: { refreshModel: true },
-    endDate: { refreshModel: true },
     filter: { refreshModel: true },
     page: { refreshModel: true },
     size: { refreshModel: true },
@@ -22,30 +16,15 @@ export default class OrganenMandatarissenRout extends Route {
     const parentModel = this.modelFor('organen.orgaan');
     const bestuursOrgaan = parentModel.bestuursorgaan;
 
-    const tijdsspecialisaties = await bestuursOrgaan.heeftTijdsspecialisaties;
-    const currentBestuursOrgaanWithPeriods =
-      getSelectedBestuursorgaanWithPeriods(tijdsspecialisaties, {
-        startDate: params.startDate,
-        endDate: params.endDate,
-      });
-
-    const bestuursPeriods = getBestuursPeriodsOG(tijdsspecialisaties);
-    const selectedPeriod = {
-      startDate: currentBestuursOrgaanWithPeriods.startDate,
-      endDate: currentBestuursOrgaanWithPeriods.endDate,
-    };
-    const currentBestuursOrgaan =
-      currentBestuursOrgaanWithPeriods.bestuursOrgaan;
-
-    const options = this.getOptions(params, currentBestuursOrgaan);
+    const options = this.getOptions(params, parentModel.currentBestuursOrgaan);
 
     const mandatarissen = await this.store.query('mandataris', options);
 
     return {
-      bestuursPeriods,
-      selectedPeriod,
       mandatarissen,
       bestuursOrgaan,
+      bestuursPeriods: parentModel.bestuursPeriods,
+      selectedPeriod: parentModel.selectedPeriod,
     };
   }
